[fix #12901] moved testSqlInjectionSearchAccountRequestsInWholeSystem to AccountRequestSearchIT class
